+++ cmd/internal/cli/singularity.go
@@ -658,7 +658,3 @@ func getBuilderClientConfig(uri string) (*scsbuildclient.Config, error) {
 
 	return currentRemoteEndpoint.BuilderClientConfig(uri)
 }
-
-func URI() string {
-	return "https://" + strings.TrimSuffix(currentRemoteEndpoint.URI, "/")
-}